Project: reject with warning if location change request can't be handled
If location changes are disabled, only the error message string has been
returned as rejection. This is now changed to a CATMAID.Warning object,
which is handled by the error handler in a less intrusive way than an
error (or a simple string).

Also, a rejection is no properly caught by the caller.

See catmaid/CATMAID#1809