Widget help: allow widgets to include documentation from HTML page
If widgets specify a `helpPath` field, their help content will be
included through an <iframe> showing the file named like what is
specified as `helpPath` in in the folder `static/html/doc/widgets/`.
This is expected to be a regular HTML file.

CATMAID will inject two style sheets (screen.css and
keyboard-shortcuts.css) and the contribution footer.

See catmaid/CATMAID#1520