Provide a workflow to use widget docs in CATMAID and in main docs
This process can be simplified further (by automating some parts), but
works already as a useful implementation. This change integrates widget
specific help pages that are used in the front-end to be included in the
main documentation as well. To do so, during a release, all widget help
pages are copied from CATMAID's static/html/widgets/ folder to the
_static/widgets/ folder in the sphinx documentation. A similar thing is
done with the API documentation. A Sphinx wrapper for each widget in the
widgets/ folder makes those HTML pages available to Sphinx and they will
be included in the list of widget help pages.

This is shown for the Selection Table.

To copy a current version of all widget docs, `make widgetdoc` can be
used. This is also what is done by the release script.

Fixes catmaid/CATMAID#1520