Tracing overlay: make LOD mode configurable through layer controls
If a grid cache is in use, a level of detail can be configured. By
default this value is computed based on the current zoom level (inverse
of zoom level relative to zoom range). This can be changed to be an
absolue zoom level, which is useful some times too.