3D viewer: add Z plane options to change background color
This works by thresholding the sum of all channels in the data for each
pixel. If it is between a configurable range [min,max], it will be
replaced by a configurable background color. This happens on the fly in
a shader.

Thanks to Helena Jambor for the suggestion of this feature.