Neuron dendrogram: fix reload after split/merge corner case
If the selected node is part of the skeleton that is not displayed in
the Neuron Dendrogram, an error is raised when trying to highlight this
node. This is fixed now. The widget will reload the active skeleton
after a split or merge. If there is an active dendrogram node, its
current skeleton ID is requested and that skeleton is reloaded.

Fixes catmaid/CATMAID#1891