Neuron similarity: add option to import similarity matrix from CSV file
The button "Create from CSV file" in the Configurations tab allows to
import a CSV file containing a similarity matrix. The format is
described in the widget documentation.

An example CSV file with a similarity matrix for the adult Drosophila
brain has been added to the static/data/neuron-similarity folder.