3D viewer: add toggle for radius display
The checkbox "Show radius" in the View Settings tab will no dynamically
control radius visibility. If radius information is hidden, the
respective meshes will be removed from the scene (or not loaded at all
initially).

This is especially useful when loading a large set of skeletonized
segmentation neurons, which typically come with radius information.
Thanks to @schlegelp for the suggestion!