DB: fix type of treenode/connector references in change_request table
Both `treenode` and `connector` are tables that inherit from `location`,
which means they share a 64-Bit ID space. The types for the respective
reference columns in the `change_request` table are however of type
`integer` and therefore can't reference a lot of treenode IDs, namely
those outsite the 32-Bit range. This is fixed by this commit and the
reference columns in the `change_request` table now also have the
correct `bigint` type.