CLI importer: add --auto-name-unknown-users option
This option will cause the importer to automatically create names for
users that are unknown, i.e. referenced in an import, but not available
from the import. Users will only be created automatically or
interactively if the --map-users or --map-user-ids options aren't used
or weren't able to map a user.