Node filter: fix intersection with volume filter rules
Node filters currently mark allowed nodes in a result object by
assigning `true` to the node ID. This is done for consistency with
Arbor.js even though there are probably faster alternatives (e.g. use
Map or Set). This didn't work correctly for volumes, because not `true`
was assigned, but `undefined` due to a wrong array lookup. In
consequence this caused volume filters to not be usable in filter rule
intersections. Thanks to Javier Valdes Aleman for pointing out this
issue.