Spatial update notifications: only try to update DB config if not in recovery
Database replicas are read-only copies of the database and are basically
constantly in recovery mode. In this state it is not possible to modify
schema or data and the spatial update notification initialization tries
to do exactly that by updating a function. This is now prevented in case
the database is in recovery mode, which makes the setup of replicas
easier.