Update id column of treenode_connector_edge and connector_geom table
This migration rewrites the tables treenode_connector_edge and
connector_geom, because their references to the treenode_connector table
are wrong and need to be updated. Up to this migration the id column was
of type int, but since it references the treenode_edge id column and the
connector id column, respectively, and they should follow their type,
which is bigint. This is updated by this migration by creating a new
table. This has the benefit that cleaning up the old table is easier and
doesn't need a VACUUM FULL run to reclaim space.

Both The treenode_connector_edge table and the connector_geom table
aren't tracked by the history system, which makes this update simpler.
In principle it could also be merged into the previous 64 Bit concept ID
migration, but it is easier for some existing test systems to not be
re-migrated.

The changes to the connector tables have been merged alrady separately
into the dev branch in commit 65b4044c4f66, but the migration didn't
allow to avoid VACUUM FULL to reclaim space, therefore the existing
version is updated accordingly. The existing version rewrote the table
directly, which would only allow to reclaim space with VACCUUM FULL.
Since this isn't very practical on large tables with many connectors,
this new version is introduced. The schema outcome is the same and
therefore no action has to be taken, if the previous version has been
applied already.

Additionally, since we recently upgrade to Django 2.2, we need to add
actual Django ORM models for models that have foreign keys, which is now
the case for connector_geom and treenode_connector_edge. Initial
performance tests showed no problem with this. The reason is that for
transaction test cases (e.g. the migration 18 test) the database will be
reset to its orginal content through TRUNCATE. Which tables are exactly
truncated is determined by the list of Django models for each app. Since
this migration defines foreign keys on project and connector and
project, transaction tests would fail, because the Django issues
TRUNCATE wouldn't include those tables before, leading to an error.