CLI exporter: add option to constrain set of exported annotations
The new --annotation-annotation option allows to specify a
meta-annotation that all exported annotations need to share
transitively.

If for instance all all top-level annotations that should be exported
are annotated with "exportable" and "--annotation-annotation exportable"
is used, only annotations from those annotation hierarchies marked as
"exportable" will be exported. Without this option, all annotations are
exported.