CLI exporter: fix corner case of constrained annotation export
Allowed annotations are specified through a meta-annotation. Currently,
we don't export this meta-annotation by default and in this case no
links of this annotation to other annotations should be exported either.
So far they have been exported though, which resulted in an consistency
error when importing such an export. This is fixed by this commit, those
class_instance_class_instance links aren't exported anymore.