Neatest node query: enforce less distance tests with skeleton filter
By using a CTE the planner is forced to compute the result separately
(until Postgres 12, where default CTE behavior changes). This guarantees
that there will be only so many distance tests (<<-->> operator) as
there are skeleton edges. Before, the planner didn't constrain by
skeleton (if passed in) when using the join directly or as subquery.

Thanks to Marta Costa (@mmc46), who noticed this performance regression
when selecting larger skeletons in the latest release.