NPM: make all dependencies dev-dependencies
We don't use NPM for runtime dependencies at the moment. The API docs
are also generated explicitly as part of the release process. This is
also done in an attempt to convince GitHub that its security alerts are
only affecting dependencies of our build process and don't imply runtime
problems.

This seems to also update many things in package-lock.json and it is not
entirely clear to me why this is the case. No version or repository
changed.