Patch PIXI.js to disable failIfMajorPerformanceCaveat option
To test if WebGL is available, PIXI.js tries to obtain a WebGL context
after it verified that the browser API is available in the first place.
Obtaining a context can be parameterized in such a way that it will fail
if the browser expects performance problems or incompatibilities. This
has been observed to happen with the Nouveau driver with Chrome on some
systems and others encountered problems also with Firefox.

This patch changes PIXI.js' WebGL test so that
failIfMajorPerformanceCaveat is set to false, which should allow the
creation of a WebGL context in more situation, potentially leading to
other problems though. Given that setting this option to false is the
default also for THREE.js, the new behavior makes is more consistent.

Plus, Nouveau implements 2D acceleration fully for Nvidia hardware, and
this is likely all that is needed from PIXI.js for rendering tiles.

If the PIXI.js version in use isn't 4.8.8 (the current version we use),
a warning is printed to the console to check this patch against the new
version.

See catmaid/CATMAID#1902