Connectivity matrix: don't attempt to create table w/o neurons
Having "Fractions" mode enabled and recreating the table without neurons
being present resulted in an error. The fraction rendering code assumed
a relation map to be loaded. While this should be fixed as well, more
work is needed to fix this properly. Until then this workaround prevents
this error. It should be fine to skip the attempt of rendering the table
without neurons.