Doc: use redoc to generate the API documentation for catmaid.org
The resulting documentation looks much nicer than what we used before
(bootprint). It also allows us to remove bootprint, which isn't actively
developed anymore and caused a lot of security alerts on GitHub that are
unlikely to get fixed.

The new static API documentation can be better linked to and integrates
a search. To make it easier to use, it is now also opened in a separate
tab.