Tracing tool: joining two skeletons w/o confirmation adds reference annotation
Joining two skeletons without confirmation ("fast mode"), will now
attach an annotation to the resulting skeleton that references the
merged in skeleton by name. The name of this annotation as well as the
suggested reference visible in the confirmation dialog have the form
"Merged: <neuron-name>".

Single node neurons are merged without such an added annotation.