Startup: use advisory locks to serialize helper function creation
When enabling and disabling both history tracking and spatial update
events on startup, advisory locks are now used. This should make the
startup of parallel CATMAID workers more robust. In some PgBouncer based
setups, I saw the following error in the database log before, which can
be fixed with this:

  PL/pgSQL function enable_spatial_update_events() line 3 at SQL statement
  2019-08-21 00:37:10.896 EDT [12385] catmaid_tracing_user@catmaid_v14_seg_li_190805_0 STATEMENT:  SELECT enable_spatial_update_events()
  2019-08-21 00:37:10.897 EDT [12401] catmaid_tracing_user@catmaid_v14_seg_li_190805_0 ERROR:  tuple concurrently updated
  2019-08-21 00:37:10.897 EDT [12401] catmaid_tracing_user@catmaid_v14_seg_li_190805_0 CONTEXT:  SQL statement "CREATE OR REPLACE FUNCTION notify_conditionally(channel text, payload text) RETURNS int
  LANGUAGE plpgsql AS
  $inner$
  BEGIN
      PERFORM pg_notify(channel, payload);
      RETURN 0;
  END;
  $inner$"

The lock ID used is formed from the multiplication of all as ascii codes
of the characters of "catmaid" + an incremental number.