Doc: fix wrong uWSGI subfolder mounter point
Thanks to @freshpomelo for pointing this out. There can be no trailing
slash in the uWSGI configuration or the CATMAID instance will misbehave.
The wrong trailing slash was added by me in 16d40f2 as a reaction to
the discussion in catmaid/CATMAID#1714, which in fact clearly showed
that there should not be a trailing slash after the mountpoint.

See catmaid/CATMAID#1912