Add parent ID to treenode_edge table
This can improve spatial query time significantly, because it allows us
to avoid the JOIN that found the parent nodes before.

The application of migration 82 might take a while to complete, because it
rewrites a potentially big table (treenode_edge). Therefore, make also sure
that there is enough space available at the database storage location (25% of
database data directory should be plenty). If no replication is used, setting
the following Postgres options can speed up the process by allowing
Postgres to skip the WAL for the migration: `wal_level = minimal`,
`archive_mode = off` and `max_wal_senders = 0`.

The table is completely recreated from scratch, because it had to be
rewritten anyway and doing so ourselves through a separate table allows
us to reclaim space more quickly and without VACUUM FULL. Indices had to
be rewritten, too, in most cases. Reindexing and constraint checking is
therefore done in bulk after the new table is created, which speeds up
the process as well.

The column alignment is optimal, and eight bytes more are now used per
tuple (one extra bigint), which makes it 105 in total.