Import/export widget: add option to import from other CATMAID instances
This only adds the basic implementation, which doesn't cooperate with
(landmark) transformations yet and doesn't track provenance of imported
data.

The new 'Import from CATMAID' tab in the Import/export widget allows to
select a remote CATMAID instance and remote project, in which it can
then search for neurons by name and annotation. Resulting neurons can be
previewed and imported in bulk or individually. The original name is
kept, but no additional data like annotations and tags are imported at
the moment.

When clicking on the "Open remotely" for a skeleton, the remote CATMAID
instance will be opened with the respective skeleton selected.