Importer: add basic provenance tracking
This is optional at the moment, but if the parameters source_id and
source_field are provided for an SWC import, a mapping between this ID
and the newly created import ID are stored in a new table
(skeleton_origin). It is stored along with a reference to a data source,
which mainly consists of a URL at the moment.