Import/export widget: show if imported skeleton has been seen before
This relies on a form of explicit provenance tracking. If the IDs of the
objects about to be imported have been seen with the current source URL
in the past, the local skeleton ID is displayed and made selectable.

The history table of the skeleton_origin table isn't used yet and
therefore, remote skeletons become 'unknown' after a merge. This will be
fixed separately.