Graph widget: add basic support for node filters
For simplicity, this is currently implemented as a subgroup mode. Only
nodes that are allowed by the selected filters, can contribute
connections and the graph. At the moment, this comes with a few caveats:
if filters are in use, no subgraphs (e.g. axon/dendrite, synapse
clusters) can be used. Grouping isn't available either and in
addition, skeletons are removed from the widget if they don't comply
with the filters. All this is likely going to change in future
versions, but this version is useful for simple queries already.