Use database default values for creation_time and edition_time
This will make the default values for these time fields consistent with
the catmaid_transaction_info entries. So far the Django default
`timezone.now` have been used, which unfortunately doesn't match the
transaction starttime (now()) that is used to keep track of each
transaction in catmaid_transaction_info. At some point Postgres will
start to reuse old transaction IDs and to keep our stored transaction ID
values unique when we lookup transactions in catmaid_transaction_info,
we need to include a time field as well. This is unlikely to happen
soon, but this change pepares our database models for that. It requires
one additional databse roundtrip though to get the current transaction
start time, which is also very quick.