Neuron navigator: add two sorting modes for neurons of annotation
This allows to use the two sorting options 'annotated_on' and
'last_annotation_link_edit'. The first one sorts by the date the matched
annotations were added to the target. The second option sorts by when
this link was edited last. This can be used for instance from an
annotation detail page in the Neuron Navigator to sort the annotated
neurons by the date they were annotated.