Navigator: prevent LastPass from filling out X/Y coordinate fields
By giving both the X and the Y input field of the Navigator toolbar the
type "search", rather than "text", LastPass doesn't seem to try and
modify these fields anymore. The "search" type behaves practically the
same as "text".

With LastPass adding itself to these fields before, URLs to views
in CATMAID that were stored in LastPass, would sometimes modify the
position of other deep links. This happened because LastPass would
update those fields after having loaded with what it stored for the last
URL to be "helpful" and autocomplete these fields. Apparently that is a
larger problem that many people run into with their web applications.
LastPass doesn't seem to care though, the issue is long known.

It doesn't seem to be a problem for dynamically added fields. The X/Y
fields are mainly affected, because they are already present in the
initial HTML.

Thanks to Marta Costa for providing additional debug information!

Fixes catmaid/CATMAID#1900