3D viewer: limit line width to 25 px max
If increased to larger values and volumetric lines are enabled, it is
easy to crash the 3D Viewer and even the browser tab, because its memory
use seems to increase significantly. To prevent this from happening when
this value is accidentally changed, an upper limit of 25 px is now
enforced. Larger values are rarely visually pleasing and commonly used
values are way lower.

Fixes catmaid/CATMAID#1899