Skeleton splitting: improve sampler handling
If a skeleton sampler was in use on a skeleton, a split could so far
only be successfully executed if the split node is within a sampler
domain of that sampler. If the splitting treenode was outside of a
domain, an error was raised. This is fixed now and sampled skeletons can
now be split upstream or downstream of a domain as well. If split
uptream, the domain will be deleted, because it is part of the split-off
fragment. This is consistent with handling the cut off part if the split
treenode is within the sampler.