Return status 400 and 500 errors on back-end errors
Ths changes CATMAID's error handling middleware so that it returns
returns either status 400, 500 or custom error status code, depending on
the exception. If the exception is an instance of the new `ClientError`
type, it is expected to have `status_code` field, which is set to 400 by
default. Permission related errors override this to return status 403.

If an error is not an instance of `ClientError`, but still a
`ValueError, status 400 is returned, because it is assumemd that a
ValueError is the result of some bad/wrong input data. Otherwise status
500 is returned.

Also fix a skeleton API test that was broken since 69c806c44f, but the
effects where masked.

Fixes catmaid/CATMAID#1921