Data stores: don't write back settings for anonymous user by default
This can be configured for each settings loading or per settings
instance, if the back-end would ever allow anonymous user to store data.
For now this fixes an error in our UI tests and gets rid of an
unnecessary and failing PUT request for storing migrated settings for
the anonymous user.