Upgrade Django from v2.1.12 to v2.2.5
Apart from a small change to the INSTALLED_APPS setting, nothing really
needed to change for this update. This update is done mainly for its
support of psycopg2 2.8 and in preparation for v3 later this year. More
details on the update are available from here:

  https://docs.djangoproject.com/en/2.2/releases/2.2/

The sqlparse module has been removed as a dependency, because it is now
an explicit dependency of Django. We don't refer to it explicitly.