Fix POST based API token login after introducing error status code
If the `accounts/login` endpoint was used to get a user's profile with a
POST request, authenticated using an API token, the view function would
raise an InvalidLoginError, because the expected user information wasn't
part of the POST request. To be consistent with the GET request, the
POST request branch will now also check for an authenticated user in the
request object. This fixes the returned error.

Also, the returned session information won't include the session key
anymore. Its remaining uses in the front end have been updated as well.
The message handling code (init.js) now allows also anonymous users to
receive messages and the Review widget now checks for authenticated
users rather than only the `id` field in the session context to test if
the current user is allowed load and save review white lists.

Debugged and written with @clbarnes.