Visibility groups: allow generic "self" creator ID test
Visibility groups can now use a generic node creator test that
references the current user (logged in our anonymous). It is called
"(self)" in the list of creators for all three visibility groups. This
is useful to define visibility groups on the project or instance level
that include a hiding pattern while guaranteeing that users see their
own nodes.