Add Django model for treenode_edge table
This is might be needed in the future in transaction tests. To save some
debugging time in the future, should there ever be a transaction test
case that uses treenode edge information, this should prevent running
into the same problems like I did for connector_geom and
treenode_connector_edge.

It should be okay to include this Django signal in an older migration,
because it is only passively there for Django to know the model has been
registered. Also, an ORDER BY clause for an update is removed, because
it doesn't have affect.