Fix spurious LastPass autocomplete in fields
This fixes LastPass and other password managers spuriously placing
autocomplete UI in most CATMAID fields, including the quick nav. This
was due to the login password field not being inside a form element,
which caused these managers to match the entire DOM as being a
password-containing form.

Related to catmaid/CATMAID#1900.