Import/export widget: persist set of default annotations for imports
A custom set of default annotations that will be added to imported
skeletons can be defined through the "Default skeleton import
annotations" option in the Settings Widget. At the moment, this is set
to "Import" as well as "{source} upload {group}". The second annotation
uses two optional placeholders, which are replaced by the skeleton
import source reference and the user's primary group, respectively.
Should no primary group be set, the username name is used.