Tracing layer: add basic remote skeleton merging
Shift + Click on a remote node from a local node (or vice versa) will
now bring up an import preview dialog which is used as a merge preview
at the same time. Confirming the change will import the remote skeleton
and merge it at the selected location, works with virtual nodes.