Submitter: execute functions w/o HTTP requests asynchronously too
A submitter can be used to queue both HTTP requests and regular
functions. HTTP requests are executed asynchronously, but so far, when
running a queue element using next(), regular functions where
executed right away in a synchronous fashion. This inconsistency can
lead to subtle problems in execution order, mainly when "exiting" the
submitter queue by generating an independent promise using "promise()".
In that case an actual Promise instance is returned that resolves as
soon as the submitter queue is at the insertion point. If however no
real asynchronous tasks where performed up to that point, a new
a new non-http continuation added to the submitter, will run *before*
the previously generated promise. This is counterintuitive and not
intended. The reason is that all continuations have been run
synchronously and the promise execution waits for the next available
async microtask slot. This commit runs therefore also sync continuations
in an async promise.

The Tracing Overlay front-end tests revealed that when working with
queueing both synchronous and asynchronous queue continuations.