3D viewer: add basic infrastructure for showing additional planes
This is already useful on the command line for some visualizations. With
3D Viewer active:

  var w = CATMAID.front();

One can now create e.g. three stack sized planes, all orthogonal to each
other:

  w.space.staticContent.createOrthoPlane(w.space, 'x', project.focusedStackViewer.primaryStack, CATMAID.Stack.ORIENTATION_XY, {color: 0xff0000, opacity: 0.5}, {color: 0xffff00, opacity: 0.7}, {x: 1.0, y: 1.0, z: 1.0});
  w.space.staticContent.createOrthoPlane(w.space, 'y', project.focusedStackViewer.primaryStack, CATMAID.Stack.ORIENTATION_XZ, {color: 0x00ff00, opacity: 0.5}, {color: 0x00ffff, opacity: 0.7}, {x: 1.0, y: 1.0, z: 1.0});
  w.space.staticContent.createOrthoPlane(w.space, 'z', project.focusedStackViewer.primaryStack, CATMAID.Stack.ORIENTATION_ZY, {color: 0x0000ff, opacity: 0.5}, {color: 0xff00ff, opacity: 0.7}, {x: 1.0, y: 1.0, z: 1.0});

The second parameter needs to be unique for each plane. Otherwise one
can't clear the space again like this:

  w.space.staticContent.removeAllOrthoPlanes(w.space)

The first parameter after the orientation is the plane color/opacity,
followed by the border color opacity. If they are null, the respective
part won't be drawn. The last parameter is a scale factor, which can be
used to make the planes smaller so that they don't cover the full extent
of the stack. This however also has the consequence that that planes can
reach outside of the stack bounding box when the location is changed.
Generally, these planes move with project location changes. You can
prevent this by providing an additional prarameter with the fixed
coordinate value for the respective depth dimension of the provide
orientation.