3D viewer: improve handling of intersecting transparent objects
Having multiple (partially) intersecting transparent objects loaded in
the 3D Viewer did often lead to surprising results. Parts of objects
that were occluded by another transparent object wouldn't be visible,
etc. This is fixed for the most part with this commit by rendering a
second version of each volume, but together with other interacting
objects like ortho planes or a regular Z plane.

Ortho planes are rendered in a separate rendering step so that
transparent objects behind them are visible. Rendering all planes in the
same rendering step leads unfortunately still to the problem that there
are views where one can't see a transparent plane behind another
transparent plane. This was also a problem before and needs to be fixed
separately.