Docker: fix parsing of some environment variable parameters
When parameter variables contained structures that could be interpreted
by Bash, they in fact were when passing them as argument to the sanitize
function, because the parameters weren't quoted. An example of this is
CM_CSRF_TRUSTED_ORIGINS, which is expected to contain a Python list (as
string):

  CM_CSRF_TRUSTED_ORIGINS="['a.example.com', 'b.int.example.com']"

This would lead to the following being written to settings.py:

  ['a.example.com',

Which clearly isn't valid Python, causing the Container start to fail.
This should be fixed by quoting the parameters to the sanitize function.