History tracking: make update/delete trigger more efficient
This database migration modifies the trigger setup of the history
tracking system of CATMAID. If history tracking is enabled, so far any
UPDATE and DELETE statement would execute a trigger function for every
touched row that would insert its old version into the table's history
table. This trigger function would also update the modified row's
transaction ID with a reference to the current transaction.

The latter part hast to happen for each row, because Postgres will not
allow us to modify updated rows (to store the transaction ID) in a
statement-level trigger. However, the first part, moving the old data to
the history table can happen on a per-statement basis, replacing N
INSERT statements with a single on. In larger setups this makes a
significant difference.

As an example, consider a well tuned server with most the database's
active set sitting in memory and a treenode table with ~800,000,000
treenodes in which we split off a fragment of ~9,000 nodes from a bigger
skeleton. So far the break-down of trigger behavior looked like this:.

 Trigger on_change_update_history_regular: time=953.769 calls=9215
 Trigger on_edit_treenode: time=35.184 calls=9215
 Trigger on_edit_treenode_check_review: time=19.663 calls=9215
 Trigger on_edit_treenode_update_summary_and_edges: time=2515.228 calls=1
 Trigger on_edit_treenode_update_suppressed_virtual_treenodes: time=17.984 calls=9215
 Trigger on_edit_treenode_update_treenode_connector_edges: time=19.190 calls=9215
 Execution Time: 4254.615 m

With the separation of the trigger behavior into two functions each, the
breakdown looks like this:

 Trigger on_change_update_history_regular: time=206.256 calls=1
 Trigger on_change_update_history_txid_regular: time=44.531 calls=9215
 Trigger on_edit_treenode: time=25.508 calls=9215
 Trigger on_edit_treenode_check_review: time=20.152 calls=9215
 Trigger on_edit_treenode_update_summary_and_edges: time=2529.471 calls=1
 Trigger on_edit_treenode_update_suppressed_virtual_treenodes: time=18.586 calls=9215
 Trigger on_edit_treenode_update_treenode_connector_edges: time=19.419 calls=9215
 Execution Time: 3410.433 ms

The reduces the history related changes from ~954ms to ~206ms + ~45ms,
which is a clear improvement. The breakdown also makes clear that the
bigger problems is the statement-level summary and edge update trigger,
which will be looked at separately.

It should be okay to just enable history tracking at the end of the new
migration. Upon a restart, CATMAID will enable/disable history tracking
according to the settings.