Store nr. of imported nodes in summary and optimize edge/summary update
The skeleton summary table contains now the number imported nodes. It is
updated by the treenode edge/summary update trigger function. This
happens in a manner that minimizes the impact on operations on skeletons
that don't contain any imported nodes (common case).

In addition, some minor aspects of this trigger function have been
improved: On the update or deletion of one or more treenode database
rows, CATMAID will update the skeleton summary tables and PostGIS edge
representation through trigger functions. These functions rely heavily
on so called transition tables that reflect the state before and after a
change of (possibly many) treenodes. It turns out some joins can be made
faster by setting more explicit limits when scanning a transition table.
Assuming new_treenode is a transition table, this means it is faster to
do

  SELECT *
  FROM treenode t
  JOIN (
    SELECT * from new_treenode
    LIMIT (SELECT COUNT(*) FROM new_treenode)
  ) nt
    ON nt.id = t.id;

than doing

  SELECT *
  FROM treenode t
  JOIN new_treenode nt
    ON nt.id = t.id;

Ideally, Postgres could infer such information in the second variant as
well. It can't at the moment though. I have used this technique already
in other parts of the main update trigger function, but didn't apply it
everywhere. This is mainly an improvement for queries that don't yield
any result rows. This improves the row estimates in many cases and can
improve the query plan. On an optimized production system with
~40.000.000 nodes, this improves the merge of a 30k skeleton into
another 30k skeleton by ~2 seconds from 19.5 to 17.3 seconds.