Improve back-end permission checks for imported treenodes
When skeletons are imported, their original creator/editor data (incl.
timestamps) is copied over, if available. To not require a change of
this, CATMAID's back-end permission checks are updated so that the
importer has full permissions on the data they imported. This allows the
importing user to modify or even delete the just imported data.

Without this, permissions on imported data would be locally
inconsistent, because not every source has creator/editor information
available. It also is what users would expect: the imported data is
handled like their own, but original information is retained.

In order to do that, the `can_edit_or_fail` family of functions check
now as last option for an approval if the requesting user happens to be
the one who imported the data. If this is the case, edits are allowed.

The front-end will now request (in parallel, non-blocking) the original
import user when a node is selected. This query is fast and shouldn't be
a problem in most cases. With slowed down network connection, this didn't
contribute to tracing action times.

Tests are included. I couldn't find any noticeable effect on query time,
even for large skeleton imports.