Tracing tool: fix selection of virtual nodes
This problem was introduced when making the Tracing Tool query import
user information (in a non-blocking way) on active node change. This
information can't be retrieved directly for virtual nodes (because they
are not real). When a virtual node is selected, their child node is now
used for the query.