Morphology plot: add interpolation options
So far the plotted data was always interpolated with a basis spline.
This isn't necessarily what is required for a particular plot, e.g. if
the sample values have to be intersected. To allow this, a drop-down
menu with the following interpolation options has been added: basis
spline, linear, cardinal spline, monotone, step before and step after.